+++ test/svelte3/integration/components/components.import.aliace.svelte
@@ -1,0 +1,4 @@
+<script>
+    /** The nested component. */
+    import { Nested as NestedWidget } from './components.nested.svelte';
+</script>