+++ test/svelte3/integration/data/data.import.aliace.svelte
@@ -1,0 +1,8 @@
+<script>
+    /**
+     * The import comment.
+     */
+    import {y as altY} from "./importable.js";
+
+    import X from "./importable.js";
+</script>